# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /read/recommend:
    post:
      operationId: getReadRecommendation
      tags: [entries]
      summary: Get AI mood-based book or short read recommendations
      description: Recommends 1-2 books or articles based on user's current mood using Gemini
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadRecommendationInput"
      responses:
        "200":
          description: AI recommendation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadRecommendationResponse"
...
components:
  schemas:
...
    ReadRecommendationInput:
      type: object
      properties:
        mood:
          type: string
          minLength: 1
          maxLength: 300
        mode:
          type: string
          enum: [books, alternatives]
      required: [mood]
    ReadRecommendationResponse:
      type: object
      properties:
        mood:
          type: string
        mode:
          type: string
        recommendedTitles:
          type: array
          items:
            type: string
        reason:
          type: string
        isCached:
          type: boolean
      required: [mood, mode, recommendedTitles, reason]
...