# lib/api-zod/src/generated/types/readRecommendationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReadRecommendationInputMode } from './readRecommendationInputMode';

export interface ReadRecommendationInput {
  /**
     * @minLength 1
     * @maxLength 300
     */
  mood: string;
  mode?: ReadRecommendationInputMode;
}