# lib/api-zod/src/generated/types/readRecommendationResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReadRecommendationResponse {
  mood: string;
  mode: string;
  recommendedTitles: string[];
  reason: string;
  isCached?: boolean;
}